pyproject.toml
[project]
name = "rag-lite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.12.0",
    "docx2txt>=0.9",
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "langchain>=1.2.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-deepseek>=1.0.1",
    "langchain-huggingface>=1.2.0",
    "langchain-milvus>=0.3.3",
    "langchain-openai>=1.1.7",
    "minio>=7.2.20",
    "pymysql>=1.1.2",
    "pypdf>=6.6.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "sqlalchemy>=2.0.45",
]
